refactor(structure): colocate files
