+++ .storybook/main.js
@@ -1,10 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  "stories": [
-    "../stories/**/*.stories.mdx",
-    "../stories/**/*.stories.@(js|jsx|ts|tsx)"
-  ],
+  "stories": ["../src/**/*.story.tsx"],
   "staticDirs": ['../public'],
   "addons": [
     "@storybook/addon-links",
