+++ .storybook/preview.js
@@ -1,4 +1,4 @@
-import '../styles/globals.css';
+import '../src/styles/globals.css';
 import * as NextImage from 'next/image';
 
 const OriginalNextImage = NextImage.default;
