+++ src/pages/artists/dedupe/components/list/__stories__/ArtistList.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { ArtistList, Skeleton } from "components/artists/dedupe/list/ArtistList"
+import { ArtistList, Skeleton } from "../ArtistList"
 
 export default {
   component: ArtistList,
