+++ src/pages/artists/dedupe/components/manager/ArtistCard.tsx
