+++ src/pages/artists/dedupe/components/manager/ArtistCardHeader.tsx
