+++ src/pages/artists/dedupe/components/manager/Confirm.tsx
