+++ src/pages/artists/dedupe/components/manager/__stories__/ArtistDupesManager.stories.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+import { ArtistDupesManager } from "../ArtistDupesManager"
+import artistWithDupes from "../__fixtures__/artist-with-dupes"
+
+export default {
+  component: ArtistDupesManager,
+}
+
+export const With3Dupes = () => <ArtistDupesManager artist={artistWithDupes} />
