+++ src/pages/artists/dedupe/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react"
 import { useMetaphysics } from "lib/artsy-next-auth"
-import { ArtistList, Skeleton } from "components/artists/dedupe/list/ArtistList"
+import { ArtistList, Skeleton } from "./components/list/ArtistList"
 
 const PER_PAGE = 36
