+++ src/pages/palette/__stories__/Palette.stories.tsx
